+++ source/Public/Get-SqlDscPreferredModule.ps1
@@ -5,6 +5,10 @@
     .DESCRIPTION
         Get the first available (preferred) module that is installed.
 
+        If the environment variable `SMODefaultModuleName` is set to a module name
+        that name will be used as the preferred module name instead of the default
+        module 'SqlServer'.
+
     .PARAMETER Name
         Specifies the list of the (preferred) modules to search for, in order.
         Defaults to 'SqlServer' and then 'SQLPS'.
@@ -51,12 +55,24 @@ function Get-SqlDscPreferredModule
     (
         [Parameter()]
         [System.String[]]
-        $Name = @('SqlServer', 'SQLPS'),
+        $Name,
 
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $Refresh
     )
+
+    if (-not $PSBoundParameters.ContainsKey('Name'))
+    {
+        if ($env:SMODefaultModuleName)
+        {
+            $Name = @($env:SMODefaultModuleName, 'SQLPS')
+        }
+        else
+        {
+            $Name = @('SqlServer', 'SQLPS')
+        }
+    }
 
     if ($Refresh.IsPresent)
     {