+++ tests/Unit/SqlServerDsc.Common.Tests.ps1
@@ -2364,6 +2364,8 @@ Describe 'SqlServerDsc.Common\Connect-SQL' -Tag 'ConnectSql' {
                         Add-Member -MemberType NoteProperty -Name ConnectAsUserPassword -Value '' -PassThru |
                         Add-Member -MemberType NoteProperty -Name ConnectAsUserName -Value '' -PassThru |
                         Add-Member -MemberType NoteProperty -Name StatementTimeout -Value 600 -PassThru |
+                        Add-Member -MemberType NoteProperty -Name ConnectTimeout -Value 600 -PassThru |
+                        Add-Member -MemberType NoteProperty -Name EncryptConnection -Value $false -PassThru |
                         Add-Member -MemberType NoteProperty -Name ApplicationName -Value 'SqlServerDsc' -PassThru |
                         Add-Member -MemberType ScriptMethod -Name Disconnect -Value {
                             return $true
@@ -2602,6 +2604,25 @@ Describe 'SqlServerDsc.Common\Connect-SQL' -Tag 'ConnectSql' {
         }
     }
 
+    Context 'When using encryption' {
+        BeforeAll {
+            Mock -CommandName New-Object `
+                -MockWith $mockNewObject_MicrosoftDatabaseEngine `
+                -ParameterFilter $mockNewObject_MicrosoftDatabaseEngine_ParameterFilter
+        }
+
+        It 'Should return the correct service instance' {
+            $mockExpectedDatabaseEngineServer = 'SERVER'
+            $mockExpectedDatabaseEngineInstance = 'SqlInstance'
+
+            $databaseEngineServerObject = Connect-SQL -Encrypt -ServerName $mockExpectedDatabaseEngineServer -InstanceName $mockExpectedDatabaseEngineInstance -ErrorAction 'Stop'
+            $databaseEngineServerObject.ConnectionContext.ServerInstance | Should -BeExactly "$mockExpectedDatabaseEngineServer\$mockExpectedDatabaseEngineInstance"
+
+            Should -Invoke -CommandName New-Object -Exactly -Times 1 -Scope It `
+                -ParameterFilter $mockNewObject_MicrosoftDatabaseEngine_ParameterFilter
+        }
+    }
+
     Context 'When connecting to the default instance using the correct service instance but does not return a correct Database Engine object' {
         Context 'When using ErrorAction set to Stop' {
             BeforeAll {
@@ -2622,6 +2643,7 @@ Describe 'SqlServerDsc.Common\Connect-SQL' -Tag 'ConnectSql' {
                                 Add-Member -MemberType NoteProperty -Name ConnectAsUserPassword -Value '' -PassThru |
                                 Add-Member -MemberType NoteProperty -Name ConnectAsUserName -Value '' -PassThru |
                                 Add-Member -MemberType NoteProperty -Name StatementTimeout -Value 600 -PassThru |
+                                Add-Member -MemberType NoteProperty -Name ConnectTimeout -Value 600 -PassThru |
                                 Add-Member -MemberType NoteProperty -Name ApplicationName -Value 'SqlServerDsc' -PassThru |
                                 Add-Member -MemberType ScriptMethod -Name Disconnect -Value {
                                     return $true
@@ -2668,6 +2690,7 @@ Describe 'SqlServerDsc.Common\Connect-SQL' -Tag 'ConnectSql' {
                                 Add-Member -MemberType NoteProperty -Name ConnectAsUserPassword -Value '' -PassThru |
                                 Add-Member -MemberType NoteProperty -Name ConnectAsUserName -Value '' -PassThru |
                                 Add-Member -MemberType NoteProperty -Name StatementTimeout -Value 600 -PassThru |
+                                Add-Member -MemberType NoteProperty -Name ConnectTimeout -Value 600 -PassThru |
                                 Add-Member -MemberType NoteProperty -Name ApplicationName -Value 'SqlServerDsc' -PassThru |
                                 Add-Member -MemberType ScriptMethod -Name Disconnect -Value {
                                     return $true